requirements.txt
# huggingface
transformers>=4.40.0,<5.0.0
datasets>=2.20.0,<3.0.0
tokenizers>=0.19.0,<0.20.0
accelerate>=0.30.0,<0.35.0
peft>=0.8.0,<0.12.0
torch>=2.3.0,<3.0.0

# dl utils
sentencepiece>=0.2.0,<0.3.0
deepspeed>=0.14.0,<0.15.0
flash-attn>=2.6.0.post1,<3.0.0; sys_platform != "darwin"

# for metrics
evaluate>=0.4.0,<0.5.0
scipy
scikit-learn
lm-eval>=0.4.3,<0.5.0
# lm-eval[vllm]>=0.4.3,<0.5.0; sys_platform != "darwin"
# lm-eval>=0.4.3,<0.5.0; sys_platform == "darwin"
calflops>=0.3.0,<0.4.0

# utils
python-dotenv
click>=8.0.0,<9.0.0
typer-slim>=0.12.0,<0.13.0
neptune>=1.10.0,<2.0.0
torchinfo>=1.8.0,<2.0.0
fvcore>=0.1.5,<0.2.0
torch_pruning==1.4.0
seaborn>=0.13.0,<0.14.0
tabulate>=0.9.0,<1.0.0
psutil>=5.0.0,<7.0.0  # for memory profiling in accelerate/transfomers

# distutils was removed in python 3.12, but used in transformers
setuptools; python_version>="3.12"
